% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCV.R
\name{bootstrapCV}
\alias{bootstrapCV}
\title{Bootstrap cross-validation}
\usage{
bootstrapCV(Y,mtr,n,...)
}
\arguments{
\item{Y}{predicted variable.}

\item{X}{set of predictors.}

\item{mtr}{mtry parameter.}

\item{nrep}{number of iteration.}
}
\value{
RMSEP data frame.
}
\description{
Evaluation of root mean square error of prediction (RMSEP) for training and validation folds
}
\examples{
bootstrapCV(IMAT,metric,2,100)
}
