% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCV2.R
\name{bootstrapCV2}
\alias{bootstrapCV2}
\title{Bootstrap cross-validation with bias correlation (70\%,30\%)}
\usage{
bootstrapCV(response,predictors,mtr,nrep)
}
\arguments{
\item{response}{predicted variable.}

\item{predictors}{set of predictors.}

\item{mtr}{mtr response parameter.}

\item{nrep}{number of iteration.}

\item{cor.method}{character, "spearman" or "pearson"}
}
\value{
RMSEP data frame.
}
\description{
Evaluation of root mean square error of prediction (RMSEP) for training and validation folds
}
