% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMaturityData.R
\name{cleanMaturityData}
\alias{cleanMaturityData}
\title{Clean database including Protest data}
\usage{
cleanMaturityData(terrain,metrics)
}
\arguments{
\item{terrain}{data frame or matrix}

\item{metrics}{data frame or matrix}
}
\value{
Return maturity database
}
\description{
Correct plot names, remove duplicated plots and select variables that are predefined
}
\examples{
# Not run (tables not available in package): maturity.db=cleanMaturityData(terrain,metrics)
}
