% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMaturityData2.R
\name{cleanMaturityData2}
\alias{cleanMaturityData2}
\title{Clean database including Protest data}
\usage{
cleanMaturityData(terrain,metrics,var.selec,var.names)
}
\arguments{
\item{terrain}{data frame or matrix}

\item{metrics}{data frame or matrix}

\item{var.selec}{character, names of the variables to keep in the database}

\item{var.names}{character, to rename variables (optional)}
}
\value{
Return maturity database
}
\description{
Correct plot names, remove duplicated plots and select defined predictor variables
}
\examples{
# Not run (tables not available in package): maturity.db=cleanMaturityData(terrain,metrics,var.selec,var.names)
}
