% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMaturityData3.R
\name{cleanMaturityData3}
\alias{cleanMaturityData3}
\title{Clean database including with terrain data selection}
\usage{
cleanMaturityData3(terrain,metrics,terrain.selec,var.selec,var.names)
}
\arguments{
\item{terrain}{data frame or matrix}

\item{metrics}{data frame or matrix}

\item{terrain.selec}{character, names of the terrain variables to keep in the database}

\item{var.selec}{character, names of the structural variables to keep in the database}

\item{var.names}{character, to rename variables (optional)}
}
\value{
Return maturity database
}
\description{
Correct plot names, remove duplicated plots and select defined predictor variables
}
\examples{
# Not run (tables not available in package): maturity.db=cleanMaturityData3(terrain,metrics,terrain.selec,var.selec,var.names)
}
