% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineStandExtraMasks2.R
\name{combineStandExtraMasks2}
\alias{combineStandExtraMasks2}
\title{Combine stand and extrapolation filters}
\usage{
combineStandExtraMasks(modelName)
}
\arguments{
\item{modelName}{character used to define output names}
}
\description{
Spatial prediction without extrapolation and for selected stands
}
\examples{
# not run: combineStandExtraMasks("IMAT")
}
