% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMaturityIndices.R
\name{computeMaturityIndices}
\alias{computeMaturityIndices}
\title{Compute maturity indices}
\usage{
computeMaturityIndices(maturityAttributeData,attribute.selec,quant)
}
\arguments{
\item{maturityAttributeData}{dataframe including maturity attributes}

\item{attribute.selec}{Colnames of maturity attributes}

\item{quant}{quantile used as a threshold to compute IMAT.}
}
\value{
Return maturity index and attributes.
}
\description{
Compute maturity indices
}
\examples{
computeMaturityIndices(maturityAttributeData="terrain",attribute.selec,quant=0.99)
}
