% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropForest.R
\name{cropForest}
\alias{cropForest}
\title{Crop forest to lidar survey extents}
\usage{
cropForest(dir)
}
\arguments{
\item{dir}{character, directory name to load BD FORET layers}
}
\description{
Crop forest to lidar survey extents and dissolve using "CODE_TFV" attribute
}
\examples{
# not run: cropForest()
}
