% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCompleteCleanData.R
\name{loadCompleteCleanData}
\alias{loadCompleteCleanData}
\title{Load and clean a dataset with all variables that were tested to model maturity.}
\usage{
loadCompleteCleanData()
}
\arguments{
\item{deleteGTGBna}{logical.}
}
\value{
Return LiDAR metric and terrain releve data frames
}
\description{
Function to load and clean data: select maturity, LiDAR and topographic variables and normalize intensity metrics. Parcel area and distance to nearest road or forest track are also included, such as LiDAR flight and observed stand type.
}
\examples{
terrain=loadCompleteCleanData(deleteGTGBna = F)[[1]]; metric=loadCompleteCleanData(deleteGTGBna = F)[[2]]
}
