% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMaturityModel.R
\name{loadMaturityModel}
\alias{loadMaturityModel}
\title{Load maturity model}
\usage{
loadMaturityModel(modelName)
}
\arguments{
\item{modelName}{character, "initModel" referred to he calibrated model from the initial dataset (660 plots), metrics calculated within a 35m radius, to make a prediction in 60m side pixels.}
}
\value{
Return the random forest model.
}
\description{
Load maturity and maturity attribute models
}
\examples{
loadMaturityModel(modelName="initModel")
}
