% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturity2.R
\name{predictMaturity2}
\alias{predictMaturity2}
\title{Predict maturity from raw metric maps}
\usage{
predictMaturity(model,modelName,dir)
}
\arguments{
\item{model}{predictive model}

\item{outputModelName}{character used to define output names}

\item{metricsDir}{character, directory name to load metric maps}

\item{metricsDataName}{character, names of metric datasets}
}
\description{
Spatial prediction without filtering
}
\examples{
# not run: impt.replicates.ord=maturitymodelling::predictMaturity2(model,outputModelName,metricsDir,metricsDataName)
}
