% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturityExtra.R
\name{predictMaturityExtra}
\alias{predictMaturityExtra}
\title{Predict maturity from raw metric maps without any extrapolation}
\usage{
predictMaturityExtra(model,modelName,dir)
}
\arguments{
\item{model}{predictive model}

\item{modelName}{character used to define output names}

\item{dir}{character, directory name to load metric maps}
}
\description{
Filter metrics to predict NA in case of extrapolation
}
\examples{
# not run: predictMaturityExtra(model,modelName,dir)
}
