% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturityStand.R
\name{predictMaturityStand}
\alias{predictMaturityStand}
\title{Filter maturity prediction according to stand composition}
\usage{
predictMaturityStand(modelName)
}
\arguments{
\item{modelName}{character used to define output names}
}
\description{
Spatial prediction with stand mask
}
\examples{
# not run: predictMaturityStand("IMAT")
}
