% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturityStand2.R
\name{predictMaturityStand2}
\alias{predictMaturityStand2}
\title{Filter maturity prediction according to a pre-defined stand composition}
\usage{
predictMaturityStand(modelName)
}
\arguments{
\item{model}{predictive model}

\item{outputModelName}{character used to define output names}

\item{metricsDir}{character, directory name to load metric maps}

\item{metricsDataName}{character, names of metric datasets}

\item{forestDir}{character, directory name to load forest shapefiles}

\item{forestShape}{character, names of forest shapefiles, must be ordered according to metricsDataName}
}
\description{
Spatial prediction with stand mask
}
\examples{
# not run: maturitymodelling::predictMaturityStand2(model,outputModelName,metricsDir,metricsDataName,forestDir,forestShape)
}
