% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilizeImportance.R
\name{stabilizeImportance}
\alias{stabilizeImportance}
\title{Stabilize conditional importance values}
\usage{
stabilizeImportance(impt.init,nrep)
}
\arguments{
\item{response}{vector, pedicted variable}

\item{impt.init}{output of the function varImpPlot used to include the "true" importance related to the initial calibrated RF model}

\item{nrep}{integer, number of replicates}
}
\value{
Return importance data frame
}
\description{
Compute conditional importance with replicates
}
\examples{
# not run: impt.replicates.ord=maturitymodelling::stabilizeImportance(impt.init,nrep)
}
