% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeIntensity.R
\name{standardizeIntensity}
\alias{standardizeIntensity}
\title{Standardize intensity by LiDAR flight}
\usage{
standardizeIntensity(maturity.db)
}
\arguments{
\item{maturity.db}{data frame or matrix}
}
\value{
Return maturity database
}
\description{
Standardization and export of mean and sd values
}
\examples{
# Not run (tables not available in package): maturity.db.stand=standardizeIntensity(maturity.db)
}
