% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCV.R
\name{bootstrapCV}
\alias{bootstrapCV}
\title{Bootstrap cross-validation (70\%, 30\%)}
\usage{
bootstrapCV(response,predictors,mtr,nrep)
}
\arguments{
\item{response}{predicted variable.}

\item{predictors}{set of predictors.}

\item{mtr}{mtr response parameter.}

\item{nrep}{number of iteration.}
}
\value{
RMSEP data frame.
}
\description{
Evaluation of root mean square error of prediction (RMSEP) for training and validation folds
}
