% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIMAT.R
\name{computeIMAT}
\alias{computeIMAT}
\title{Compute maturity indices}
\usage{
computeIMAT(useGTTGB,diameter,quant)
}
\arguments{
\item{useGTTGB}{logical, use GTGBGTTGB if set to false.}

\item{diameter}{20cm or 30cm threshold for standing and lying deadwood.}

\item{quant}{quantile used as a threshold to compute IMAT.}
}
\value{
Return maturity index and attributes.
}
\description{
Compute the maturity index described in Furh et al. (in prep.) and maturity attributes (living tree basal area, standing deadwood basal area, lying deadwood volume).
}
\examples{
computeIMAT(useGTTGB=T,diameter=30,quant=0.99)
}
