% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCleanData.R
\name{loadCleanData}
\alias{loadCleanData}
\title{Load and clean data used in Fuhr et al. (in prep.); private data access.}
\usage{
loadCleanData()
}
\arguments{
\item{deleteGTGBna}{logical.}
}
\value{
Return LiDAR metric and terrain releve data frames
}
\description{
Function to load and clean data: select maturity, LiDAR and topographic variables and normalize intensity metrics.
}
\examples{
terrain=loadCleanData(deleteGTGBna = F)[[1]]; metric=loadCleanData(deleteGTGBna = F)[[2]]
}
