% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtryOptimization.R
\name{mtryOptimization}
\alias{mtryOptimization}
\title{mtry optimization}
\usage{
mtryOptimization(Y,X,nrep,...)
}
\arguments{
\item{Y}{predicted variable.}

\item{X}{set of predictors.}

\item{nrep}{number of iteration.}
}
\value{
Plot the mtry value that minimize the mean out-of-bag error rate.
}
\description{
Function to define mtry random forest parameter.
}
\examples{
mtryOptimization(IMAT,metric,100)
}
