% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturity.R
\name{predictMaturity}
\alias{predictMaturity}
\title{Predict maturity from raw metric maps}
\usage{
predictMaturity(model,modelName,dir)
}
\arguments{
\item{model}{predictive model}

\item{modelName}{character used to define output names}

\item{dir}{character, directory name to load metric maps}
}
\description{
Spatial prediction without filtering
}
\examples{
# not run: impt.replicates.ord=maturitymodelling::predictMaturity(model=protestModel_IMAT,modelName="IMAT",dir=dir)
}
