% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaturityForest.R
\name{predictMaturityForest}
\alias{predictMaturityForest}
\title{Predict maturity from raw metric maps}
\usage{
predictMaturity(modelName)
}
\arguments{
\item{modelName}{character used to define output names}
}
\description{
Spatial prediction with forest mask
}
\examples{
# not run: predictMaturityForest("IMAT")
}
